Person Tester application
An abstract Person class that stores first name, last name, and email
address. A class named Customer inherits the Person class and it stores
a customer number. Employee inherits the Person class and it stores the
social security number. The Main class creates the Customer and
Employee objects from the data entered by the user, and it uses these
objects to display the data to the user.